funcs_consumer/pyproject.toml
[project]
name = "funcs-servicer"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "customtkinter>=5.2.2",
    "dotenv>=0.9.9",
    "pika>=1.3.2",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
]
